requirements.txt
gradio
groq
PyPDF2
python-docx
python-dotenv